Move logicalfile, physicalfile out of __init__.py
